requirements.txt
alembic==1.4.2
aniso8601==8.0.0
argon2-cffi==20.1.0
attrs==19.3.0
autopep8==1.5.2
blinker==1.4
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
flake8==3.8.2
flasgger==0.9.4
Flask==1.1.2
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.3.0
packaging==20.4
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.8.1
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.4.2
pytest-mock==3.1.0
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
six==1.15.0
SQLAlchemy==1.3.17
urllib3==1.25.9
wcwidth==0.1.9
Werkzeug==1.0.1
